requirements.txt
alabaster==0.7.12
asgiref==3.3.1
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-crispy-forms==1.11.2
docutils==0.16
idna==2.10
imagesize==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.1.2
pyenchant==3.2.0
Pygments==2.8.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
snowballstemmer==2.1.0
sqlparse==0.4.1
urllib3==1.26.4
